terraMind-backend/requirements.txt
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Mail==0.9.1
python-dotenv==0.19.2
mysql-connector-python==8.0.27
requests==2.26.0
openai==0.27.0
Flask-Cors==3.1.1
Flask-Login==0.6.2
